feat: refactor typography imports and create Typography component
