+++ src/frontend/todos-web/src/app/examples/greeting/page.tsx
@@ -1,5 +1,5 @@
 import { AppPage } from "@/components/layout/app-page";
-import { TypographyH1 } from "@/components/typography/typography-h1";
+import { TypographyH1 } from "@/components/typography/typography";
 import {
   GreetingCardConfig,
   GreetingClient,
