+++ src/frontend/todos-web/src/app/examples/shadcn/page.tsx
@@ -1,5 +1,5 @@
 import { AppPage } from "@/components/layout/app-page";
-import { TypographyH1 } from "@/components/typography/typography-h1";
+import { TypographyH1 } from "@/components/typography/typography";
 import { Button } from "@/components/ui/button";
 
 export default function ShadcnPage() {
